DjangoRestApisPostgreSQL/requirements.txt
Django==3.1.6
djangorestframework==3.12.2
django-cors-headers==3.7.0
psycopg2-binary==2.8.5
